fix: ssc structure
