+++ src/app/[lang]/exercises/cc/[slug]/introduction/page.tsx
@@ -2,9 +2,21 @@ import {About, Button, NavBar, RiveAnimation} from '@/components';
 import {ccMock} from '@/lib/mock';
 import {ArrowRight} from 'lucide-react';
 import Link from 'next/link';
+import {useMemo} from 'react';
 
-export default function IntroductionPage({params}: {params: {lang: string; slug: string}}) {
+export default function CCIntroductionPage({params}: {params: {lang: string; slug: string}}) {
   const {slug} = params;
 
-  return <About {...(slug === 'check-in' ? ccMock.checkIn.about : ccMock.checkOut.about)} />;
+  const data = useMemo(() => {
+    switch (slug) {
+      case 'check-in':
+        return ccMock.checkIn.about;
+      case 'check-out':
+        return ccMock.checkOut.about;
+      default:
+        return ccMock.checkIn.about;
+    }
+  }, [slug]);
+
+  return <About {...data} />;
 }
